Prefilter and msw changes

+++ rWorkspace/msw_cluster/MswAlgorithm.R
@@ -23,7 +23,7 @@ CalculateMSWMatrix <-  function(input_file_path ,
     #   return (FALSE)
     # }
     data_set = data_set %>% na.omit()
-    alarm_seq = paste0(data_set$Alarm.Tag, data_set$Alarm.Id)
+    alarm_seq = paste0(data_set[,5], data_set[,6])
 	
     # Writing csv data to text files
     write.table(
@@ -46,14 +46,30 @@ CalculateMSWMatrix <-  function(input_file_path ,
   #distance matrix for hierarchical clustering
   scores_matrix = as.data.frame.matrix(scores)
   d <- dist(scores_matrix, method = "euclidean")
-  #saving similarity matrix in csv
+  #merge similarity index along with prefilter values 
   setwd(similarity_index_path)
-  write.csv(
+  res_file <- paste0(similarity_index_path,"/","result", ".csv") 
+  if(file.exists(res_file)){
+   write.table(
     scores_frame,
     file = paste0("result", ".csv"),
-    row.names = FALSE,
-    quote = FALSE
-  )
+    append = TRUE,
+    sep=',',
+    quote = FALSE,
+    col.names = FALSE,
+    row.names=FALSE
+    )
+  }
+  else
+    {
+      write.csv(
+        scores_frame,
+        file = paste0("result", ".csv"),
+        row.names = FALSE,
+        quote = FALSE
+      )
+    }
+  }
   return(scores_frame)
 }
